Created library geometry.
