+++ src/openMVG/cameras/CMakeLists.txt
@@ -1,14 +1,14 @@
 
-UNIT_TEST(openMVG Camera_IO "openMVG_multiview;stlplus")
+UNIT_TEST(openMVG Camera_IO "openMVG_multiview;stlplus;openMVG_geometry")
 
-UNIT_TEST(openMVG Camera_Pinhole "openMVG_multiview")
+UNIT_TEST(openMVG Camera_Pinhole "openMVG_multiview;openMVG_geometry")
 
-UNIT_TEST(openMVG Camera_Pinhole_Radial "openMVG_multiview")
+UNIT_TEST(openMVG Camera_Pinhole_Radial "openMVG_multiview;openMVG_geometry")
 
-UNIT_TEST(openMVG Camera_Pinhole_Brown "openMVG_multiview")
+UNIT_TEST(openMVG Camera_Pinhole_Brown "openMVG_multiview;openMVG_geometry")
 
-UNIT_TEST(openMVG Camera_Pinhole_Fisheye "openMVG_multiview")
+UNIT_TEST(openMVG Camera_Pinhole_Fisheye "openMVG_multiview;openMVG_geometry")
 
-UNIT_TEST(openMVG Camera_Spherical "openMVG_multiview")
+UNIT_TEST(openMVG Camera_Spherical "openMVG_multiview;openMVG_geometry")
 
-UNIT_TEST(openMVG Camera_Subset_Parametrization "openMVG_multiview")
+UNIT_TEST(openMVG Camera_Subset_Parametrization "openMVG_multiview;openMVG_geometry")
