+++ src/openMVG/cameras/Camera_Pinhole.hpp
@@ -15,6 +15,7 @@
 #include "openMVG/cameras/Camera_Intrinsics.hpp"
 #include "openMVG/geometry/pose3.hpp"
 #include "openMVG/numeric/eigen_alias_definition.hpp"
+#include "openMVG/multiview/projection.hpp"
 
 namespace openMVG
 {
