+++ src/openMVG/geometry/rigid_transformation3D_srt.hpp
@@ -9,7 +9,6 @@
 #ifndef OPENMVG_GEOMETRY_RIGID_TRANSFORMATION_3D_SRT_HPP
 #define OPENMVG_GEOMETRY_RIGID_TRANSFORMATION_3D_SRT_HPP
 
-#include <limits>
 #include "openMVG/numeric/lm.hpp"
 #include "openMVG/numeric/eigen_alias_definition.hpp"
 
@@ -35,52 +34,14 @@ namespace geometry
  *
  * \note Need at least 3 points
  */
-inline bool FindRTS
-(
-  const Mat &x1,
-  const Mat &x2,
-  double * S,
-  Vec3 * t,
-  Mat3 * R
-)
-{
-  if ( x1.cols() < 3 || x2.cols() < 3 )
-  {
-    return false;
-  }
-
-  assert( 3 == x1.rows() );
-  assert( 3 <= x1.cols() );
-  assert( x1.rows() == x2.rows() );
-  assert( x1.cols() == x2.cols() );
-
-  // Get the transformation via Umeyama's least squares algorithm. This returns
-  // a matrix of the form:
-  // [ s * R t]
-  // [ 0 1]
-  // from which we can extract the scale, rotation, and translation.
-  const Eigen::Matrix4d transform = Eigen::umeyama( x1, x2, true );
-
-  // Check critical cases
-  *R = transform.topLeftCorner<3, 3>();
-  if ( R->determinant() < 0 )
-  {
-    return false;
-  }
-  *S = pow( R->determinant(), 1.0 / 3.0 );
-  // Check for degenerate case (if all points have the same value...)
-  if ( *S < std::numeric_limits<double>::epsilon() )
-  {
-    return false;
-  }
-
-  // Extract transformation parameters
-  *S = pow( R->determinant(), 1.0 / 3.0 );
-  *R /= *S;
-  *t = transform.topRightCorner<3, 1>();
-
-  return true;
-}
+  bool FindRTS
+  (
+    const Mat &x1,
+    const Mat &x2,
+    double * S,
+    Vec3 * t,
+    Mat3 * R
+  );
 
 /**
 * @brief Eigen Levemberg-Marquardt functor to refine translation, Rotation and Scale parameter.
@@ -99,43 +60,14 @@ struct lm_SRTRefine_functor : Functor<double>
   */
   lm_SRTRefine_functor( int inputs, int values,
                         const Mat &x1, const Mat &x2,
-                        const double &S, const Mat3 & R, const Vec &t ): Functor<double>( inputs, values ),
-    x1_( x1 ), x2_( x2 ), t_( t ), R_( R ), S_( S ) { }
+                        const double &S, const Mat3 & R, const Vec &t );
 
   /**
   * @brief Computes error given a sample
   * @param x a Sample
   * @param[out] fvec Error for each values
   */
-  int operator()( const Vec &x, Vec &fvec ) const
-  {
-    // convert x to rotation matrix and a translation vector and a Scale factor
-    // x = {tx,ty,tz,anglex,angley,anglez,S}
-    Vec3 transAdd = x.block<3, 1>( 0, 0 );
-    Vec3 rot = x.block<3, 1>( 3, 0 );
-    double Sadd = x( 6 );
-
-    //Build the rotation matrix
-    Mat3 Rcor =
-      ( Eigen::AngleAxis<double>( rot( 0 ), Vec3::UnitX() )
-        * Eigen::AngleAxis<double>( rot( 1 ), Vec3::UnitY() )
-        * Eigen::AngleAxis<double>( rot( 2 ), Vec3::UnitZ() ) ).toRotationMatrix();
-
-    const Mat3 nR  = R_ * Rcor;
-    const Vec3 nt = t_ + transAdd;
-    const double nS = S_ + Sadd;
-
-    // Evaluate re-projection errors
-    Vec3 proj;
-    for ( Mat::Index i = 0; i < x1_.cols(); ++i )
-    {
-      proj = x2_.col( i ) -  ( nS *  nR * ( x1_.col( i ) ) + nt );
-      fvec[i * 3]   = proj( 0 );
-      fvec[i * 3 + 1] = proj( 1 );
-      fvec[i * 3 + 2] = proj( 2 );
-    }
-    return 0;
-  }
+  int operator()( const Vec &x, Vec &fvec ) const;
 
   Mat x1_, x2_;
   Vec3 t_;
@@ -161,41 +93,14 @@ struct lm_RRefine_functor : Functor<double>
   */
   lm_RRefine_functor( int inputs, int values,
                       const Mat &x1, const Mat &x2,
-                      const double &S, const Mat3 & R, const Vec &t ): Functor<double>( inputs, values ),
-    x1_( x1 ), x2_( x2 ), t_( t ), R_( R ), S_( S ) { }
+                      const double &S, const Mat3 & R, const Vec &t );
 
   /**
    * @brief Computes error given a sample
    * @param x a Sample
    * @param[out] fvec Error for each values
    */
-  int operator()( const Vec &x, Vec &fvec ) const
-  {
-    // convert x to rotation matrix
-    // x = {anglex,angley,anglez}
-    Vec3 rot = x.block<3, 1>( 0, 0 );
-
-    //Build the rotation matrix
-    Mat3 Rcor =
-      ( Eigen::AngleAxis<double>( rot( 0 ), Vec3::UnitX() )
-        * Eigen::AngleAxis<double>( rot( 1 ), Vec3::UnitY() )
-        * Eigen::AngleAxis<double>( rot( 2 ), Vec3::UnitZ() ) ).toRotationMatrix();
-
-    const Mat3 nR  = R_ * Rcor;
-    const Vec3 nt = t_;
-    const double nS = S_;
-
-    // Evaluate re-projection errors
-    Vec3 proj;
-    for ( Mat::Index i = 0; i < x1_.cols(); ++i )
-    {
-      proj = x2_.col( i ) -  ( nS *  nR * ( x1_.col( i ) ) + nt );
-      fvec[i * 3]   = proj( 0 );
-      fvec[i * 3 + 1] = proj( 1 );
-      fvec[i * 3 + 2] = proj( 2 );
-    }
-    return 0;
-  }
+  int operator()( const Vec &x, Vec &fvec ) const;
 
   Mat x1_, x2_;
   Vec3 t_;
@@ -215,64 +120,14 @@ struct lm_RRefine_functor : Functor<double>
  *
  * \return none
  */
-inline void Refine_RTS
+void Refine_RTS
 (
   const Mat &x1,
   const Mat &x2,
   double * S,
   Vec3 * t,
   Mat3 * R
-)
-{
-  {
-    lm_SRTRefine_functor functor( 7, 3 * x1.cols(), x1, x2, *S, *R, *t );
-
-    Eigen::NumericalDiff<lm_SRTRefine_functor> numDiff( functor );
-
-    Eigen::LevenbergMarquardt<Eigen::NumericalDiff<lm_SRTRefine_functor> > lm( numDiff );
-    lm.parameters.maxfev = 1000;
-    Vec xlm = Vec::Zero( 7 ); // The deviation vector {tx,ty,tz,rotX,rotY,rotZ,S}
-
-    lm.minimize( xlm );
-
-    const Vec3 transAdd = xlm.block<3, 1>( 0, 0 );
-    const Vec3 rot = xlm.block<3, 1>( 3, 0 );
-    const double SAdd = xlm( 6 );
-
-    //Build the rotation matrix
-    const Mat3 Rcor =
-      ( Eigen::AngleAxis<double>( rot( 0 ), Vec3::UnitX() )
-        * Eigen::AngleAxis<double>( rot( 1 ), Vec3::UnitY() )
-        * Eigen::AngleAxis<double>( rot( 2 ), Vec3::UnitZ() ) ).toRotationMatrix();
-
-    *R = ( *R ) * Rcor;
-    *t = ( *t ) + transAdd;
-    *S = ( *S ) + SAdd;
-  }
-
-  // Refine rotation
-  {
-    lm_RRefine_functor functor( 3, 3 * x1.cols(), x1, x2, *S, *R, *t );
-
-    Eigen::NumericalDiff<lm_RRefine_functor> numDiff( functor );
-
-    Eigen::LevenbergMarquardt<Eigen::NumericalDiff<lm_RRefine_functor> > lm( numDiff );
-    lm.parameters.maxfev = 1000;
-    Vec xlm = Vec::Zero( 3 ); // The deviation vector {rotX,rotY,rotZ}
-
-    lm.minimize( xlm );
-
-    const Vec3 rot = xlm.block<3, 1>( 0, 0 );
-
-    //Build the rotation matrix
-    const Mat3 Rcor =
-      ( Eigen::AngleAxis<double>( rot( 0 ), Vec3::UnitX() )
-        * Eigen::AngleAxis<double>( rot( 1 ), Vec3::UnitY() )
-        * Eigen::AngleAxis<double>( rot( 2 ), Vec3::UnitZ() ) ).toRotationMatrix();
-
-    *R = ( *R ) * Rcor;
-  }
-}
+);
 
 } // namespace geometry
 } // namespace openMVG
