+++ src/openMVG/sfm/CMakeLists.txt
@@ -19,6 +19,7 @@ list(REMOVE_ITEM sfm_files_header ${REMOVEFILESUNITTEST})
 add_library(openMVG_sfm ${sfm_files_header} ${sfm_files_cpp})
 target_link_libraries(
   openMVG_sfm
+  openMVG_geometry
   openMVG_multiview
   stlplus
   ${CERES_LIBRARIES}
