+++ src/openMVG_Samples/cameras_undisto_Brown/CMakeLists.txt
@@ -1,6 +1,7 @@
 
 add_executable(openMVG_sample_cameras_undistoBrown undistoBrown.cpp)
 target_link_libraries(openMVG_sample_cameras_undistoBrown
+  openMVG_geometry
   openMVG_image
   openMVG_multiview
   stlplus
